- refactoring in tests - fix server error handler registry
