+++ src/test/java/de/zalando/mass/ratwrap/bean/AnotherTestRegistryBean.java
@@ -1,0 +1,4 @@
+package de.zalando.mass.ratwrap.bean;
+
+public class AnotherTestRegistryBean {
+}
