--- src/test/java/de/zalando/mass/ratwrap/controller/AnotherTestRegistryBean.java
@@ -1,4 +1,0 @@
-package de.zalando.mass.ratwrap.controller;
-
-public class AnotherTestRegistryBean {
-}
