+++ src/test/java/de/zalando/mass/ratwrap/controller/FilteredController.java
@@ -3,6 +3,9 @@ package de.zalando.mass.ratwrap.controller;
 import de.zalando.mass.ratwrap.annotation.Filtered;
 import de.zalando.mass.ratwrap.annotation.RequestController;
 import de.zalando.mass.ratwrap.annotation.RequestHandler;
+import de.zalando.mass.ratwrap.controller.filter.TestAnnotatedFilter;
+import de.zalando.mass.ratwrap.controller.filter.TestClassAnnotatedFilter;
+import de.zalando.mass.ratwrap.controller.filter.TestNotBeanFilter;
 
 @RequestController(uri = "filtered/")
 @Filtered(TestClassAnnotatedFilter.class)
