+++ src/test/java/de/zalando/mass/ratwrap/controller/SSEStreamController.java
@@ -1,6 +1,7 @@
 package de.zalando.mass.ratwrap.controller;
 
 import com.google.common.primitives.Ints;
+import de.zalando.mass.ratwrap.Utils;
 import de.zalando.mass.ratwrap.annotation.HeaderParam;
 import de.zalando.mass.ratwrap.annotation.RequestController;
 import de.zalando.mass.ratwrap.annotation.RequestHandler;
